refactor(applicant): 인증코드 인증 서비스 인터페이스 삭제
